Skip lines when scrolling
If the delay for scrolling a single line would be too small (calculated
from `config.max_length`), then skip some lines so that the animation
would be smoother.

Currently, the `smallest_delay` (0.5ms) is just a random value that
looks okay for me (Kitty terminal). But I think that a good value for
this depends on the rendering speed of the terminal. A larger
`smallest_delay` would probably be needed for slower terminals,
otherwise, the animation would still take longer than desired.